Some cleanups and refactorings.
